+++ examples/stackdriver/outputs.tf
@@ -17,11 +17,3 @@
 output "gsuite-export" {
   value = module.gsuite-export
 }
-
-output "gsuite-log-export" {
-  value = module.gsuite-log-export
-}
-
-output "bigquery" {
-  value = module.bigquery
-}