+++ test/fixtures/bigquery/versions.tf
@@ -14,14 +14,6 @@
  * limitations under the License.
  */
 
-output "gsuite-export" {
-  value = module.gsuite-export
-}
-
-output "gsuite-log-export" {
-  value = module.gsuite-log-export
-}
-
-output "bigquery" {
-  value = module.bigquery
+terraform {
+  required_version = ">= 0.12"
 }