+++ test/setup/variables.tf
@@ -1,5 +1,5 @@
 /**
- * Copyright 2018 Google LLC
+ * Copyright 2019 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,15 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-output "gsuite-export" {
-  value = module.gsuite-export
+variable "org_id" {
+  description = "The numeric organization id"
 }
 
-output "gsuite-log-export" {
-  value = module.gsuite-log-export
+variable "folder_id" {
+  description = "The folder to deploy in"
 }
 
-output "bigquery" {
-  value = module.bigquery
+variable "billing_account" {
+  description = "The billing account id associated with the project, e.g. XXXXXX-YYYYYY-ZZZZZZ"
 }