+++ CHANGELOG.md
@@ -12,8 +12,11 @@ Extending the adopted spec, each change should have a link to its corresponding 
 
 ### Added
 - Support for Terraform 0.12
-- Kitchen-terraform tests
-- Private IP for GSuite exporter VM
+- Add Kitchen-terraform tests
+- Update examples to fit terraform-google-log-export reformat
+
+### Removed
+- Removed bats tests (obsolete)
 
 ## [0.1.0] - 2019-03-29